Move component images to their own directory
Define image dimensions inline
