--- client/src/components/Bio/style.css
@@ -1,6 +1,0 @@
-.bio .card > img {
-  width: 8rem;
-  height: 8rem;
-  display: flex;
-  flex-wrap: wrap;
-}
