+++ client/src/components/Navbar/style.css
@@ -2,12 +2,6 @@
   background-color: #fffdd0;
   box-shadow: 0 4px 2px -2px rgba(0, 0, 0, 0.2);
 }
-
-.navbar-brand img {
-  width: 50px;
-  height: 50px;
-}
-
 .nav-link {
   display: inline-block;
 }
